Correccion de epic 1 propiedades

+++ src/main/java/com/example/demo/repositorios/PersonaRepositorio.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entidades.Persona;
 
-public interface PersonaRepositorio extends JpaRepository<Persona, Integer> {
+public interface PersonaRepositorio extends JpaRepository<Persona, Long> {
 
 	List<Persona> findByEliminadoFalse();
 }
